Create 3043. Find the Length of the Longest Common Prefix
This repository contains a C++ solution for finding the longest common prefix between two arrays of integers. The solution utilizes a Trie (prefix tree) data structure to efficiently compute the longest common prefix.

The solution involves the following steps:

Trie Node Structure: A trienode struct is defined with an array of child pointers.
Insert Function: A function to insert the string representation of integers into the Trie.
Prefix Match Function: A function to find the length of the common prefix between a given string and the strings stored in the Trie.
Main Function: The longestCommonPrefix function that processes the input arrays and utilizes the Trie to find the longest common prefix.