Create 432. All O`one Data Structure
The AllOne class is a data structure that supports the following operations:

Increment (inc): Increments the count of a key by 1. If the key does not exist, it is added with a count of 1.
Decrement (dec): Decrements the count of a key by 1. If the key’s count becomes 0, it is removed from the data structure.
Get Maximum Key (getMaxKey): Returns one of the keys with the highest count.
Get Minimum Key (getMinKey): Returns one of the keys with the lowest count.
Implementation Details
The data structure uses a doubly linked list to maintain the order of counts.
Each node in the linked list contains a count and a list of keys with that count.
An unordered map (mp) is used to quickly access the node corresponding to a key.
Key Methods
addNode(node* prevNode, int cnt): Adds a new node with the given count after the specified node.
removeNode(node* curr): Removes the specified node from the linked list.
inc(string key): Increments the count of the specified key.
dec(string key): Decrements the count of the specified key.
getMaxKey(): Returns the key with the maximum count.
getMinKey(): Returns the key with the minimum count.